requirements.txt
numpy==1.24.3
pyyaml==6.0
mlflow==2.3.2
dvc==3.10.0
pytest==7.4.0
matplotlib==3.7.1